+++ classes/database/wpdk-db.php
@@ -252,7 +252,10 @@ SQL;
       foreach ( $source_row as $field => $value ) {
         $destination_object->$field = $value;
         if ( method_exists( $destination_object, 'column_' . $field ) ) {
-          call_user_func( array( $destination_object, 'column_' . $field ), $value );
+          call_user_func( array(
+                               $destination_object,
+                               'column_' . $field
+                          ), $value );
         }
       }
       return $destination_object;
@@ -264,7 +267,7 @@ SQL;
    * Return a single instance of $object class or an array of $object class. FALSE otherwise.
    * If select more rows as array, the OBJECT_K flag is used. In this way you have the key array equal to id of the country
    *
-   * @brief Select a record set
+   * @brief      Select a record set
    *
    * @param int    $id         ID of record or array of id
    * @param object $object     Object to map record fields. This object is maped when id is a single number.
@@ -276,11 +279,12 @@ SQL;
    *                           row's first column's value. Duplicate keys are discarded.
    *
    * @deprecated Use _select() instead
-   * @todo  This method should be improve with new object_query and where() method below.
+   * @todo       This method should be improve with new object_query and where() method below.
    *
    * @return object|array
    */
-  public function select( $id = false, $order_by = '', $order = 'ASC', $where = '' ) {
+  public function select( $id = false, $order_by = '', $order = 'ASC', $where = '' )
+  {
     global $wpdb;
 
     $sql_where = '';
@@ -371,7 +375,7 @@ SQL;
    *
    * @brief Build where condiction
    *
-   * @param WPDKDBTableRow $query An instance of WPDKDBTableRow
+   * @param WPDKDBTableRow $query       An instance of WPDKDBTableRow
    * @param string         $prefix      Optional. Table prefix.
    *
    * @return string
@@ -436,8 +440,9 @@ SQL;
   /**
    * @deprecated Use select() instead
    */
-  public function selectWhereID( $id, $object, $output = OBJECT ) {
-    _deprecated_function( __METHOD__, '1.0.0', 'delete()' );
+  public function selectWhereID( $id, $object, $output = OBJECT )
+  {
+    _deprecated_function( __METHOD__, '1.0.0', 'select()' );
     $this->select( $id );
   }
 
@@ -458,8 +463,8 @@ SQL;
  * @class              WPDKDBTableRow
  * @author             =undo= <info@wpxtre.me>
  * @copyright          Copyright (C) 2012-2013 wpXtreme Inc. All Rights Reserved.
- * @date               2013-05-21
- * @version            1.0.0
+ * @date               2013-09-27
+ * @version            1.0.1
  * @note               No stable - Used by SmartShop carrier
  *
  */
@@ -595,6 +600,17 @@ class WPDKDBTableRow {
     $result = $wpdb->get_results( $sql, OBJECT_K );
 
     return $result;
+  }
+
+  /**
+   * Override this method to return a filtered key pairs array with column name and default value
+   *
+   * @brief Defaults value
+   */
+  public function defaults()
+  {
+    /* Override */
+    return array();
   }
 }